desktop_analyzer/requirements.txt
PyQt6>=6.4.0
Pillow>=9.0.0
google-generativeai>=0.3.0
google-cloud-secret-manager>=2.16.0
pytest>=7.0.0
pytest-qt>=4.0.0 